ZentrumBean wurde geandert und in folge dessen musste der entsprechende JUnit Test angepasst werden. Kleine Aenderung beim PatientBean.
git-svn-id: https://randi2.svn.sourceforge.net/svnroot/randi2/randi/trunk/RANDI2@872 ccfe945e-4522-0410-b90e-900a8cb55c58

+++ src/de/randi2/model/fachklassen/beans/ZentrumBean.java
@@ -97,7 +97,7 @@ public class ZentrumBean extends Filter {
 	 * @param passwortHash
 	 *            String - Passwort bereits gehasht.
 	 */
-	public ZentrumBean(int id, String institution, String abteilung,
+	public ZentrumBean(long id, String institution, String abteilung,
 			String ort, String plz, String strasse, String hausnr,
 			long ansprechpartnerId, String passwortHash) {
